lgaleana/Refactor
