+++ server/services/modassembly/supabase/client.ts
@@ -8,3 +8,5 @@ const supabase = createClient(
   process.env.SUPABASE_URL,
   process.env.SUPABASE_SERVICE_ROLE_KEY
 );
+
+export { supabase };
