+++ server/services/modassembly/supabase/profiles.ts
@@ -1,13 +1,4 @@
-import { createClient } from '@supabase/supabase-js';
-
-if (!process.env.SUPABASE_URL || !process.env.SUPABASE_SERVICE_ROLE_KEY) {
-  throw new Error('Missing Supabase environment variables');
-}
-
-const supabase = createClient(
-  process.env.SUPABASE_URL,
-  process.env.SUPABASE_SERVICE_ROLE_KEY
-);
+import { supabase } from '@/lib/modassembly/supabase/client';
 
 /**
  * Get the Stripe customer ID for a user
